requirements.txt
Django>=3.2,<3.4
djangorestframework>=3.13,<3.14
psycopg2>=2.9,<3.0
gunicorn>=20.1,<20.2
djangorestframework-simplejwt>=5.1,<5.2
drf-yasg>=1.20,<1.25
coverage>=6.4,<6.5
# PyJWT>=2.4,<2.5
#uWSGI>=2.0,<2.1
